# dify/openapi_tools.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DDG-Agent Single-Tools",
    "description": "对公尽调单步工具：公开搜索、企业工商、司法风险",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "本地开发环境"
    }
  ],
  "paths": {
    "/tools/search": {
      "post": {
        "summary": "公开网页搜索",
        "description": "调用 Bocha 搜索引擎获取公开网页结果，自动标注来源可信度。",
        "operationId": "toolSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "搜索成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResponse"
                }
              }
            }
          },
          "502": {
            "description": "上游搜索服务异常"
          }
        }
      }
    },
    "/tools/business": {
      "post": {
        "summary": "企业工商信息查询",
        "description": "通过元典企业库查询工商基础信息（股东、法人、注册资本等）及涉诉统计。",
        "operationId": "toolBusiness",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResponse"
                }
              }
            }
          },
          "502": {
            "description": "上游企业库异常"
          }
        }
      }
    },
    "/tools/legal": {
      "post": {
        "summary": "企业司法风险查询",
        "description": "通过元典案例库检索企业相关诉讼、被执行、失信、行政处罚等司法风险。",
        "operationId": "toolLegal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResponse"
                }
              }
            }
          },
          "502": {
            "description": "上游司法库异常"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SearchRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "query": {
            "type": "string",
            "description": "搜索关键词"
          },
          "max_results": {
            "type": "integer",
            "default": 8,
            "description": "返回结果数量（1-50）"
          },
          "freshness": {
            "type": "string",
            "default": "noLimit",
            "description": "时间范围：noLimit / oneDay / oneWeek / oneMonth / oneYear"
          },
          "include": {
            "type": "string",
            "default": "",
            "description": "限定域名，多个用 | 分隔"
          },
          "exclude": {
            "type": "string",
            "default": "",
            "description": "排除域名，多个用 | 分隔"
          }
        }
      },
      "BusinessRequest": {
        "type": "object",
        "required": ["enterprise_name"],
        "properties": {
          "enterprise_name": {
            "type": "string",
            "description": "企业全称或关键词"
          }
        }
      },
      "LegalRequest": {
        "type": "object",
        "required": ["enterprise_name"],
        "properties": {
          "enterprise_name": {
            "type": "string",
            "description": "企业全称或关键词"
          },
          "top_k": {
            "type": "integer",
            "default": 10,
            "description": "返回案例数量（1-50）"
          }
        }
      },
      "ToolResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "provider": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "enterprise_name": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}